Add Guava Optional collector to GuavaCollectors. Remove duplicate Guava collector factories in tests.
