+++ sqlobject/src/test/java/org/jdbi/v3/docs/TestPaging.java
@@ -25,12 +25,12 @@ import org.hamcrest.CoreMatchers;
 import org.jdbi.v3.H2DatabaseRule;
 import org.jdbi.v3.Handle;
 import org.jdbi.v3.Something;
+import org.jdbi.v3.guava.GuavaCollectors;
 import org.jdbi.v3.sqlobject.Bind;
 import org.jdbi.v3.sqlobject.SomethingMapper;
 import org.jdbi.v3.sqlobject.SqlBatch;
 import org.jdbi.v3.sqlobject.SqlObjectBuilder;
 import org.jdbi.v3.sqlobject.SqlQuery;
-import org.jdbi.v3.sqlobject.TestCollectorFactory;
 import org.jdbi.v3.sqlobject.customizers.RegisterCollectorFactory;
 import org.jdbi.v3.sqlobject.customizers.RegisterMapper;
 import org.junit.Before;
@@ -81,7 +81,7 @@ public class TestPaging
     }
 
     @RegisterMapper(SomethingMapper.class)
-    @RegisterCollectorFactory(TestCollectorFactory.ImmutableListCollectorFactory.class)
+    @RegisterCollectorFactory(GuavaCollectors.Factory.class)
     public interface Sql
     {
         @SqlBatch("insert into something (id, name) values (:id, :name)")
